refactor: 패키지 구조 변경 - options의 클래스들을 option으로 이동
