+++ backend/src/main/java/com/h2o/h2oServer/domain/option/application/OptionService.java
@@ -2,15 +2,16 @@ package com.h2o.h2oServer.domain.option.application;
 
 import com.h2o.h2oServer.domain.option.dto.OptionDetailsDto;
 import com.h2o.h2oServer.domain.option.dto.OptionDto;
+import com.h2o.h2oServer.domain.option.dto.TrimDefaultOptionDto;
+import com.h2o.h2oServer.domain.option.dto.TrimExtraOptionDto;
 import com.h2o.h2oServer.domain.option.entity.HashTagEntity;
 import com.h2o.h2oServer.domain.option.entity.OptionDetailsEntity;
 import com.h2o.h2oServer.domain.option.entity.OptionEntity;
 import com.h2o.h2oServer.domain.option.exception.NoSuchOptionException;
 import com.h2o.h2oServer.domain.option.mapper.OptionMapper;
 import com.h2o.h2oServer.domain.optionPackage.mapper.PackageMapper;
-import com.h2o.h2oServer.domain.options.dto.*;
-import com.h2o.h2oServer.domain.options.entity.TrimDefaultOptionEntity;
-import com.h2o.h2oServer.domain.options.entity.TrimExtraOptionEntity;
+import com.h2o.h2oServer.domain.option.entity.TrimDefaultOptionEntity;
+import com.h2o.h2oServer.domain.option.entity.TrimExtraOptionEntity;
 import com.h2o.h2oServer.global.util.Validator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
