+++ backend/src/test/java/com/h2o/h2oServer/domain/option/api/OptionControllerTest.java
@@ -4,10 +4,10 @@ import com.h2o.h2oServer.domain.option.HashTagFixture;
 import com.h2o.h2oServer.domain.option.application.OptionService;
 import com.h2o.h2oServer.domain.option.dto.OptionDetailsDto;
 import com.h2o.h2oServer.domain.option.entity.enums.OptionCategory;
-import com.h2o.h2oServer.domain.options.dto.TrimDefaultOptionDto;
-import com.h2o.h2oServer.domain.options.dto.TrimExtraOptionDto;
-import com.h2o.h2oServer.domain.options.entity.TrimDefaultOptionEntity;
-import com.h2o.h2oServer.domain.options.entity.TrimExtraOptionEntity;
+import com.h2o.h2oServer.domain.option.dto.TrimDefaultOptionDto;
+import com.h2o.h2oServer.domain.option.dto.TrimExtraOptionDto;
+import com.h2o.h2oServer.domain.option.entity.TrimDefaultOptionEntity;
+import com.h2o.h2oServer.domain.option.entity.TrimExtraOptionEntity;
 import com.h2o.h2oServer.domain.trim.Exception.NoSuchTrimException;
 import com.h2o.h2oServer.global.config.RedisConfig;
 import org.junit.jupiter.api.DisplayName;
